+++ src/finlang.test/TranspilerTest/C99TranspilerTest.cs
@@ -32,4 +32,42 @@ public class C99TranspilerTest
     //    files.HasFileName("MyClass.h").Should().BeTrue();
     //    files.HasFileName("MyClass.c").Should().BeFalse();
     //}
+
+    /// <summary>
+    /// https://github.com/fin-language/fin/issues/30
+    /// </summary>
+    [Fact]
+    public void wrap_add_sub_mul_30()
+    {
+        const string finCode = """
+            using finlang;
+
+            public class MyClass: FinObj
+            {
+                static void calc()
+                {
+                    u8 a = 255;
+                    u8 b = 2;
+                    a = a.wrap_add(b);
+                    a = a.wrap_add(3*6);
+
+                    u16 c = 5;
+                    c = c.wrap_sub(10/2);
+
+                    u32 d = 256;
+                    d = d.wrap_mul(256*256);
+                }
+            }
+            """;
+
+        var files = TranspilerTestHelper.TranspileFinToCFilesWithDummyMain(finCode);
+        string cCode = files.GetSingleWriterTextByFileName("MyClass.c");
+        // addition
+        cCode.Should().Contain("a = (uint8_t)(a + (b))");
+        cCode.Should().Contain("a = (uint8_t)(a + (3*6))");
+        // subtraction
+        cCode.Should().Contain("c = (uint16_t)(c - (10/2))");
+        // multiplication
+        cCode.Should().Contain("d = (uint32_t)(d * (256*256))");
+    }
 }